% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epinion_color.R
\name{epi_cols}
\alias{epi_cols}
\title{Extract HEX codes by Epinion name of the color}
\usage{
epi_cols(...)
}
\arguments{
\item{...}{Takes the Epinion name of the color as input - i.e. Epinion Red, Dark Blue, Dark Purple, etc.}
}
\value{
A HEX code of the specified color
}
\description{
This function gives the user an easy access to the official Epinion colors by their name. This can be used every time you need to specify
a color in your code as it just returns the HEX code which in return R converts into its respective color.
}
