% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epinion_color.R
\name{epi_pal}
\alias{epi_pal}
\title{Extract a palette of Epinion colors}
\usage{
epi_pal(palette = "main", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Defaults to main palette. Other palettes include extra, extra2 and full.}

\item{reverse}{Defaults to FALSE which implies that the colors are returned in the order they appear in the palette. TRUE reverses this order.}

\item{...}{Additional arguments may apply.}
}
\value{
A palette of official Epinion colors
}
\description{
This function gives the user easy access to different palettes of Epinion colors commonly used.
}
