% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epinion_dst.R
\name{epinion_dst}
\alias{epinion_dst}
\title{Function to update population matrix from Danmarks Statistik}
\usage{
epinion_dst(variable)
}
\arguments{
\item{variable}{Argument that defines which variable you need to distribution for. Currently, the argument takes values such as "køn",
"alder", "køn_alder", "region", "uddannelse2", "person", "husstand" and "partivalg_sidste_fv".}
}
\value{
A dataframe with the variable specified and a share of the population
}
\description{
This function provides the utility of Danmark Statistik's API and cleans up the results for Epinion consultants to use in the daily
workflow with survey data and weights. The bulk of the function was written by Isak Dyrløv Klindt. Thus, all praise and errors should
be contributed to him.
}
