% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epinion_color.R
\name{epinion_fill}
\alias{epinion_fill}
\title{Apply Epinion colors to graph}
\usage{
epinion_fill(palette = "main", discrete = TRUE, reverse = FALSE, ...)
}
\arguments{
\item{palette}{Defaults to main. If you need extra colors or the full palette, please specify extra, extra2 or full.}

\item{discrete}{Defaults to TRUE. If you have an inherit continous variable, please change to FALSE.}

\item{reverse}{Defaults to FALSE. If you need to reverse the order of the colors in the palette, change this to TRUE.}

\item{...}{Additional arguments may apply.}
}
\value{
A CVI compliant plot
}
\description{
This function allows you to add Epinion colors to an existing plot by adding this function at the end of your ggplot2 syntax.
The function is especially convenient when the fill argument is specified by a variable in your data.
}
