% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epinion_style.R
\name{epinion_style}
\alias{epinion_style}
\title{Add Epinion colors to your graph}
\usage{
epinion_style(legend = FALSE)
}
\arguments{
\item{legend}{Defaults to FALSE. If you want to include a legend, please change this setting to TRUE. It will add the legend at bottom.}
}
\value{
A CVI compliant plot
}
\description{
This function can be added at the end of your ggplot2 syntax. Thus it takes a plot object as an input and adds a layer consisting of
the Epinion guidelines for conducting visualisations. In this way, it is convient to be compliant with the CVI when you are on the fly.
As a defualt it does not include a lengend if you split your data by a group variable, but this can easily be added be setting the
legend argument to TRUE.
}
